Organization
Renamed sudoku folder to the problem number (to avoid naming issues eventually as
more problems are added).  The triangle folder is not being renamed since it
involves two problems (the src txt documents have the problem number).  Added the
solution file for problem 1 (verbose but good practice) and a Makefile for easily
running the problems.